chore(payments-server): move loading overlay to directory
- refs #3718
