+++ packages/fxa-payments-server/src/components/LoadingOverlay/index.scss
