+++ packages/fxa-payments-server/src/components/LoadingOverlay/index.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { LoadingOverlay } from './LoadingOverlay';
+import { LoadingOverlay } from './index';
 
 storiesOf('components/LoadingOverlay', module).add('isLoading = true', () => (
   <LoadingOverlay isLoading={true} />
