+++ packages/fxa-payments-server/src/components/LoadingOverlay/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import { LoadingOverlay } from './LoadingOverlay';
+import { LoadingOverlay } from './index';
 
 afterEach(cleanup);
